fix: Allow catching event subscription error
Fixed #127
