+++ src/models/event-errors.ts
@@ -1,0 +1,13 @@
+export class EventsError extends Error {
+  constructor(public code: EventsErrorCode, baseError?: Error) {
+    super(baseError?.message);
+    this.name = 'EventsError';
+    this.stack = baseError?.stack;
+  }
+}
+
+export enum EventsErrorCode {
+  SubscribeFailed = 'SubscribeFailed',
+  RenewSubscriptionFailed = 'RenewSubscriptionFailed',
+  UnsubscribeFailed = 'UnsubscribeFailed',
+}
